Novo layout de artigos e temas

+++ app/Http/Controllers/admin/TemaController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\admin;
+namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use App\Models\Tema;
@@ -25,7 +25,10 @@ class TemaController extends Controller
                      ->where('titulo','LIKE','%'.$request->pesquisa.'%');
             $temas = $query->orderBy('id','DESC')->paginate(5);
         }
-        return view('tema.index',compact('temas','color'));
+        return view('tema.index',[
+            'temas' => $temas,
+            'color' => $color,
+        ]);
     }
     
     public function create()
@@ -37,30 +40,25 @@ class TemaController extends Controller
     public function store(Request $request)
     {
         $validator = Validator::make($request->all(),[
-            'titulo'     => 'required|max:100',
-            'descricao'  => 'required|max:180', 
-        ],[
-            'titulo.required'  => 'O campo TÍTULO é obrigatório',
-            'titulo.max'       => 'O TÍTULO não pode ter mais de :max caracteres!',
-            'descricao.require'=> 'O campo DESCRIÇÃO é obrigatório',
-            'descricao.max'    => 'A DESCRIÇÃO não pode ter mais de :max caracteres!',
+            'titulo'     => ['required','max:100'],
+            'descricao'  => ['required','max:180'],
         ]);
         if($validator->fails()){
             return response()->json([
                 'status' => 400,
                 'message' => $validator->errors()->getMessages(),
             ]);
-        }else{            
-            $data = [
-                'id' => $this->maxidtema_inc(),
-                'titulo'    => $request->input('titulo'),
-                'descricao' => $request->input('descricao'),
-                //'slug'      => $request->input('slug'),                
-            ];
-            $tema = $this->tema->create($data); //registro criado            
-            $t = Tema::find($tema->id);
+        }else{        
+            $data['id'] = $this->maxId();
+            $data['titulo'] = $request->input('titulo');
+            $data['descricao'] = $request->input('descricao');            
+            $data['created_at'] = now();
+            $data['updated_at'] = null;
+
+            $tema = $this->tema->create($data);
+            
             return response()->json([
-                'tema' => $t, //o objeto $t é atribuído ao json tema
+                'tema' => $tema,
                 'status' => 200,
                 'message' => 'Registro gravado com sucesso!',
             ]);
@@ -74,7 +72,7 @@ class TemaController extends Controller
     }
 
     
-    public function edit($id)
+    public function edit(int $id)
     {
         $tema = $this->tema->find($id);
         return response()->json([
@@ -84,16 +82,11 @@ class TemaController extends Controller
     }
 
     
-    public function update(Request $request, $id)
+    public function update(Request $request, int $id)
     {
         $validator = Validator::make($request->all(),[
-            'titulo'     => 'required|max:100',
-            'descricao'  => 'required|max:180', 
-        ],[
-            'titulo.required'  => 'O campo TÍTULO é obrigatório',
-            'titulo.max'       => 'O TÍTULO não pode ter mais de :max caracteres!',
-            'descricao.require'=> 'O campo DESCRIÇÃO é obrigatório',
-            'descricao.max'    => 'A DESCRIÇÃO não pode ter mais de :max caracteres!',
+            'titulo'     => ['required','max:100'],
+            'descricao'  => ['required','max:180'],
         ]);
         if($validator->fails()){
             return response()->json([
@@ -103,13 +96,11 @@ class TemaController extends Controller
         }else{
             $tema = $this->tema->find($id);            
             if($tema){
-                $data = [
-                    'titulo'    => $request->input('titulo'),
-                    'descricao' => $request->input('descricao'),
-                    //'slug'      => $request->input('slug'),
-                ];
-                $tema->update($data);
-                $t = $this->tema->find($id);
+                $data['titulo'] = $request->input('titulo');
+                $data['descricao'] = $request->input('descricao');                
+                $data['updated_at'] = now();
+                $tema->update($data); //retorna um valor booleano
+                $t = Tema::find($id); //registro atualizado
                 return response()->json([
                     'tema'    => $t,
                     'status'  => 200,
@@ -126,11 +117,17 @@ class TemaController extends Controller
     }
 
     
-    public function destroy($id)
+    public function destroy(int $id)
     {
         $tema = $this->tema->find($id);
-        $a = $tema->artigos;
-        $tema->artigos()->detach($a);
+        $artigos = $tema->artigos;
+        if($tema->artigos->count()){
+        //$tema->artigos()->detach($artigos);
+            return response()->json([
+                'status' => 400,
+                'errors' => 'Este registro não pode ser excluído! Pois há outros que dependem dele.',
+            ]);
+        }
         $tema->delete();
         return response()->json([
             'status'  => 200,
@@ -138,13 +135,15 @@ class TemaController extends Controller
         ]);
     }
 
-    protected function maxidtema_inc(){
+    protected function maxId(){
         $tema = $this->tema->orderByDesc('id')->first();
         if($tema){
-            $codigo = $tema->id+1;
+            $codigo = $tema->id;
         }else{
-            $codigo = 1;
+            $codigo = 0;
         }
-        return $codigo;
+        return $codigo+1;
     }
+
+
 }
